+++ inst/examples/xspreadsheet/app.R
@@ -1,0 +1,7 @@
+# load the required libraries
+library(RXSpreadsheet)
+library(shiny)
+library(highcharter)
+library(bslib)
+
+# TO DO: use example.R logic here ?