+++ NEWS.md
@@ -1,4 +1,9 @@
 # RXSpreadsheet 2.1.1.7
 
+* Known issues:
+- Merged cells cause a failure when saving spreadsheet state and reloading ->
+maybe the saved JSON object is incomplete wrst merged cells.
+- Filter menu does not fit if the spreadsheet height is too small
+
 * Working towards a new release.
 * Added a `NEWS.md` file to track changes to the package.